# handlers/json.go
package handlers

import (
	"encoding/json"
	"log"
	"net/http"
)

type HTTPError struct {
	Error string `json:"error" example:"Something went wrong"`
}

func RespondWithJSON(w http.ResponseWriter, code int, payload interface{}) {
	data, err := json.Marshal(payload)
	if err != nil {
		log.Printf("Failed to marshal JSON response: %v", payload)
		w.WriteHeader(500)
		return
	}

	w.Header().Add("Content-Type", "application/json")
	w.WriteHeader(code)
	w.Write(data)
}

func RespondWithError(w http.ResponseWriter, code int, msg string) {
	if code > 499 {
		log.Println("Responding with 5XX error:", msg)
	}

	RespondWithJSON(w, code, HTTPError{
		Error: msg,
	})
}